+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Drawer for menus
 - Current location data
 - Site Message
+- Forbidden, not allowed etc margin top
 
 
 ## [1.0.0] - 2019-05-19